+++ nzbhydra/stats.py
@@ -296,7 +296,7 @@ def get_nzb_downloads(page=0, limit=100, type=None):
 
 # ((Search.identifier_value == MovieIdCache.imdb) & (Search.identifier_key == "imdbid"))
 def get_search_requests(page=0, limit=100, type=None):
-    query = Search().select(Search.time, Search.internal, Search.query, Search.identifier_key, Search.identifier_value, Search.category, Search.season, Search.episode, Search.type, Search.username, TvIdCache.title.alias("tvtitle"), MovieIdCache.title.alias("movietitle")).join(TvIdCache,
+    query = Search().select(Search.time, Search.internal, Search.query, Search.identifier_key, Search.identifier_value, Search.category, Search.season, Search.episode, Search.type, Search.username, Search.title, Search.author, TvIdCache.title.alias("tvtitle"), MovieIdCache.title.alias("movietitle")).join(TvIdCache,
                                                                                                                                                                                                                                                                                      JOIN.LEFT_OUTER, on=(
             ((Search.identifier_value == TvIdCache.tvdb) & (Search.identifier_key == "tvdbid")) |
             ((Search.identifier_value == TvIdCache.tvrage) & (Search.identifier_key == "rid"))