+++ ui-src/js/search-history-controller.js
@@ -3,7 +3,7 @@ angular
     .controller('SearchHistoryController', SearchHistoryController);
 
 
-function SearchHistoryController($scope, $state, StatsService, history) {
+function SearchHistoryController($scope, $state, StatsService, history, $sce) {
     $scope.type = "All";
     $scope.limit = 100;
     $scope.pagination = {
@@ -87,7 +87,47 @@ function SearchHistoryController($scope, $state, StatsService, history) {
             return "Update query";
         }
         return request.query;
-    }
+    };
+
+    $scope.formatAdditional = function(request) {
+        var result = [];
+        //ID key: ID value
+        //season
+        //episode
+        //author
+        //title
+        if (request.identifier_key) {
+            var href;
+            var key;
+            if (request.identifier_key == "imdbid") {
+                key = "IMDB ID";
+                href = "https://www.imdb.com/title/tt"
+            } else  if (request.identifier_key == "tvdbid") {
+                key = "TVDB ID";
+                href = "https://thetvdb.com/?tab=series&id="
+            } else if (request.identifier_key == "rid") {
+                key = "TVRage ID";
+                href = "internalapi/redirect_rid?rid="
+            } else if (request.identifier_key == "tmdb") {
+                key = "TMDV ID";
+                href = "https://www.themoviedb.org/movie/"
+            }
+            result.push(key + ": " + '<a target="_blank" href="' + href + request.identifier_value + '">' + request.identifier_value + "</a>");
+        }
+        if (request.season) {
+            result.push("Season: " + request.season);
+        }
+        if (request.episode) {
+            result.push("Episode: " + request.episode);
+        }
+        if (request.author) {
+            result.push("Author: " + request.author);
+        }
+        if (request.title) {
+            result.push("Title: " + request.title);
+        }
+        return $sce.trustAsHtml(result.join(", "));
+    };
 
 
 }